// src/types/apiResponseType/OrdersResponseType.ts
import { OrderItem } from "../UserOrdersTypes";

export interface PaginationLink {
  url: string | null;
  label: string;
  active: boolean;
}

export interface OrdersResponseType {
  current_page: number;
  data: OrderItem[];

  first_page_url: string;
  from: number;
  last_page: number;
  last_page_url: string;

  links: PaginationLink[];

  next_page_url: string | null;
  path: string;
  per_page: number;
  prev_page_url: string | null;
  to: number;
  total: number;
}